Removes product ID from push interaction
Removes the `productId` property from the `NotificationModel` and updates the `sendNotification` click event to no longer require or utilize it.
Also renames `username` to `externalId` in `UserModel`.

The main reason for this change is to simplify the push notification interaction process by removing a redundant parameter. This makes the implementation cleaner and less prone to errors due to unnecessary data.